error while converting volcalendar from js to tsx
